python/requirements.txt
attrs==19.3.0
bleach==3.1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coverage==5.0.4
cryptography==2.8
docutils==0.16
entrypoints==0.3
flake8==3.7.9
googleapis-common-protos==1.51.0
grpcio==1.28.1
grpcio-tools==1.28.1
idna==2.9
importlib-metadata==1.6.0
jeepney==0.4.3
keyring==21.2.0
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
pkginfo==1.5.0.1
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
SecretStorage==3.1.2
six==1.14.0
tqdm==4.44.1
twine==3.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
zipp==3.1.0
